New commit

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -9,11 +9,12 @@
  */
 int main(void)
 {
-int mi;
+long int mi;
+
 for (mi = 0 ; mi < 10 ; mi++)
 {
         putchar(mi);
 }
-putchar("\n");
+putchar('\n');
 return (0);
 }
